requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==5.0
EbookLib==0.18
idna==3.6
lxml==4.9.3
multidict==6.0.4
Pillow==10.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
python-slugify==8.0.1
requests==2.31.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
text-unidecode==1.3
urllib3==2.1.0
yarl==1.9.2
